# Update benchmarks/targets.py: inference/forward/backward parametrization