Three main things:
* It's now possible to inspect directly with usual Thunder tools the traces of the function returned from `executor(fn)`,
* All benchmarks are now parametrized by the "compute type": inference, training forward, or backward,
* Forward and backward passes for benchmarking are now split, previously we benchmarked them together.

Let's take a look at the simplest example of benchmarking gelu function.
Previous version of the benchmark:
```py
def test_nanogpt_gelu_grad(benchmark, executor: Callable):
    gelu_bench: Benchmark = NanoGPTGeLUBenchmark(
        config="gpt2-xl", device="cuda:0", dtype=thunder.bfloat16, requires_grad=True
    )

    args, kwargs = gelu_bench.make_batch()
    # NOTE 'executor' is not simply thunder.jit meaning that thunder.last_traces(fn) won't work
    fn = executor(gelu_bench)

    benchmark(fn, *args, **kwargs)
```
new proposed in this PR:
```py
@pytest.mark.parametrize(
    "executor,",
    executors,
    ids=executors_ids,
)
@parametrize_compute_type
def test_nanogpt_gelu(benchmark, executor: Callable, compute_type: ComputeType):
    gelu_bench: Benchmark = NanoGPTGeLUBenchmark(
        config="gpt2-xl", device="cuda:0", dtype=thunder.bfloat16, requires_grad=is_requires_grad(compute_type)
    ) # No changes here

    args, kwargs = gelu_bench.make_batch() # no changes here
    # NOTE 'executor' now is simply thunder.jit
    fn = executor(gelu_bench.fn()) # <-- passing Benchmark.fn() instead of Benchmark

    benchmark_for_compute_type(compute_type, benchmark, fn, args, kwargs) # New helper function for parametrizing inference/forward/backward

    # Inserting a breakpoint after the benchmark call allows calling
    # thunder.last_traces(fn) or thunder.last_backward_traces(fn)
```
Previously it was difficult to find out what's the correct place to set the breakpoint and inspect Thunder traces. Not anymore! Try it yourself by adding a `breakpoint()` after `benchmark_for_compute_type` and call from command like
```bash
pytest thunder/benchmarks/targets.py -k "test_nanogpt_gelu[traning_backward-thunder]" --benchmark-group-by='param:compute_type' --pdb
```
Sample terminal output when run the command above without pdb:
```py

-------------------------------------------------------------------------------------- benchmark 'compute_type=ComputeType.INFERENCE': 4 tests --------------------------------------------------------------------------------------
Name (time in us)                                                  Min                 Max               Mean            StdDev             Median               IQR             Outliers  OPS (Kops/s)            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_nanogpt_gelu[inference-torch]                             19.4400 (1.0)       28.3500 (1.0)      20.1547 (1.0)      0.2564 (1.0)      20.2050 (1.0)      0.1300 (1.0)      5426;5272       49.6162 (1.0)       26796           2
test_nanogpt_gelu[inference-torch.compile]                     44.2700 (2.28)      74.5910 (2.63)     45.8282 (2.27)     1.1512 (4.49)     45.6100 (2.26)     0.6090 (4.68)     1133;1190       21.8206 (0.44)      23332           1
test_nanogpt_gelu[inference-thunder]                           79.2600 (4.08)     215.2810 (7.59)     83.1868 (4.13)     4.2078 (16.41)    81.9910 (4.06)     1.8000 (13.85)    1325;2345       12.0211 (0.24)      12502           1
test_nanogpt_gelu[inference-thunder+nvfuser+torch.compile]     79.6400 (4.10)     216.6710 (7.64)     83.4611 (4.14)     4.1994 (16.38)    82.2000 (4.07)     1.9845 (15.26)    1285;2227       11.9816 (0.24)      12476           1
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------- benchmark 'compute_type=ComputeType.TRAINING_BACKWARD': 4 tests -----------------------------------------------------------------------------------
Name (time in us)                                                  Min                 Max                Mean            StdDev              Median               IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_nanogpt_gelu[backward-torch]                              50.6900 (1.0)      112.4410 (1.0)       61.6071 (1.0)      2.5207 (1.0)       61.3800 (1.0)      1.1000 (1.0)     1812;1942       16.2319 (1.0)       23262           1
test_nanogpt_gelu[backward-torch.compile]                      72.7600 (1.44)     243.1510 (2.16)      80.6562 (1.31)     7.6725 (3.04)      77.9800 (1.27)     2.8090 (2.55)    1382;2073       12.3983 (0.76)      11819           1
test_nanogpt_gelu[backward-thunder+nvfuser+torch.compile]     108.9410 (2.15)     156.2110 (1.39)     112.9611 (1.83)     2.7181 (1.08)     112.3100 (1.83)     1.5490 (1.41)      887;842        8.8526 (0.55)       9182           1
test_nanogpt_gelu[backward-thunder]                           109.7300 (2.16)     447.9720 (3.98)     115.6489 (1.88)     5.6825 (2.25)     114.7600 (1.87)     1.6990 (1.54)     614;1154        8.6469 (0.53)       9316           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------- benchmark 'compute_type=ComputeType.TRAINING_FORWARD': 4 tests ----------------------------------------------------------------------------------
Name (time in us)                                                Min                 Max                Mean            StdDev              Median               IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_nanogpt_gelu[forward-torch]                             19.9400 (1.0)       86.6100 (1.0)       20.3519 (1.0)      0.4648 (1.0)       20.3250 (1.0)      0.1000 (1.0)      473;1132       49.1356 (1.0)       25674           2
test_nanogpt_gelu[forward-torch.compile]                     54.5100 (2.73)      91.1500 (1.05)      56.8700 (2.79)     1.6089 (3.46)      56.5700 (2.78)     1.0499 (10.50)    1124;820       17.5840 (0.36)      17993           1
test_nanogpt_gelu[forward-thunder]                           98.6500 (4.95)     237.1600 (2.74)     102.9608 (5.06)     4.1199 (8.86)     101.8105 (5.01)     2.2700 (22.70)    948;1128        9.7124 (0.20)      10046           1
test_nanogpt_gelu[forward-thunder+nvfuser+torch.compile]     99.0710 (4.97)     227.2010 (2.62)     103.0882 (5.07)     4.1031 (8.83)     101.8910 (5.01)     2.0300 (20.30)   1081;1674        9.7004 (0.20)       9997           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

cc @crcrpar @tfogal, @kevinstephano