@IvanYashchuk 2024-06-03
> Cool! Should we add the `torch.utils.benchmark.timer` timer already with this PR for this benchmarks?

This timer is used by default now. It was added in https://github.com/Lightning-AI/lightning-thunder/pull/491.